test: StubAdapter の三重定義を共通フィクスチャに集約（#28）
全抽象メソッドをスタブ実装した具象サブクラスが test_adapter_base.py の
MinimalAdapter と test_adapters/test_base.py の StubAdapter に重複定義され、
base.py のシグネチャ変更時に複数箇所の修正を要していた。

- tests/conftest.py に共通 StubAdapter（+ make_stub_adapter ファクトリ）を 1 箇所定義。
- 両テストファイルは `from tests.conftest import StubAdapter` で再利用
  （既存の `from tests.*.conftest import ...` 慣習に倣う）。
- 挙動・アサーションは不変。全 3870 passed（テスト数維持）。

Refs #28
